separando home e page: os dois componentes precisam está em arquivo diferentes para se renderizar
